refactor(engine): split tiers.ts into types, definitions, and helpers
Extracts the 578-line tier config into 3 focused modules:
- tiers/types.ts (type definitions — what most importers need)
- tiers/definitions.ts (TIERS data object, constants — data-heavy)
- tiers/helpers.ts (getTier, isValidTier, formatStorage, etc.)

Original tiers.ts becomes a thin re-export barrel — all existing
import paths continue to work unchanged. The ~30 importers that only
need `type TierKey` can now pull from types.ts (~80 lines) instead
of the full 578-line file.
